Moved data() to project context
